Updated ollama inference system
Changed the way ollama is interacted with. Before the Python api was used but now the model is hosted on the users local machine and is pinged as an API. Doing inference this way means that there is no need for a python environment in the bundle it also means that the response time is faster.